requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx>=0.26,<0.29
pydantic>=2.5.0,<3.0
python-dotenv==1.0.0
aiofiles==23.2.1
asyncpg==0.29.0
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
transformers
torch
numpy
emoji==0.6.0

# Database clients (environment-dependent)
supabase==2.17.0  # For production environment